Final cobbled-together version that works.  Demoed in YouTube video.  Plenty to fix and improve.  No shutter control, DeviceComm is really janky, etc.

+++ CamSliderCommander/SliderAxisPositionInfo.cs
@@ -8,6 +8,6 @@ namespace CamSliderCommander
 {
     public class SliderAxisPositionInfo
     {
-        public decimal Position { get; set; }
+        public double Position { get; set; }
     }
 }
